+++ src/main/java/com/x1/frans/user/command/service/UserCommandService.java
@@ -12,5 +12,5 @@ public interface UserCommandService {
 
     void createSupplierUser(SupplierUserRequestVO supplierUserRequestVO);
 
-    void changePassword(Integer userId, ChangePasswordRequestVO changePasswordRequestVO);
+    void changePassword(Long userId, ChangePasswordRequestVO changePasswordRequestVO);
 }
